requirements.txt
# Core Engine
playwright==1.42.0
pytest-playwright==0.5.2

# Test Runner & Framework
pytest==8.3.5
pytest-asyncio==0.24.0

# Environment & Configuration
python-dotenv==1.0.1

# Reporting
allure-pytest==2.13.3

# Utilities
# asyncio (included in Python 3.7+, but sometimes pinned)